zukou: client library
