+++ clients/meson.build
@@ -1,3 +1,4 @@
-client_inc = [include_directories('.'), public_inc]
+client_inc = [ include_directories('include'), public_inc ]
 
+subdir('zukou')
 subdir('simple-box')
