+++ clients/simple-box/box.h
@@ -1,34 +1,23 @@
-#ifndef GGN_CLIENT_BOX_H
-#define GGN_CLIENT_BOX_H
+#ifndef ZEN_CLIENT_SIMPLE_BOX_BOX_H
+#define ZEN_CLIENT_SIMPLE_BOX_BOX_H
 
-#include "app.h"
-#include "zgn-opengl.h"
-#include "zgn-virtual-object.h"
+#include <zukou.h>
 
-class Box
+class Box : public zukou::VirtualObject
 {
  public:
-  Box(App *app);
-  ~Box();
+  Box(zukou::App *app);
   void Frame(uint32_t time);
-  ZgnVirtualObject<Box> *virtual_object();
-  ZgnOpenGL *opengl();
+  inline zukou::OpenGLComponent *component();
 
  private:
-  ZgnVirtualObject<Box> *virtual_object_;
-  ZgnOpenGL *opengl_;
+  zukou::OpenGLComponent *component_;
 };
 
-inline ZgnVirtualObject<Box> *
-Box::virtual_object()
+inline zukou::OpenGLComponent *
+Box::component()
 {
-  return virtual_object_;
+  return component_;
 }
 
-inline ZgnOpenGL *
-Box::opengl()
-{
-  return opengl_;
-}
-
-#endif  //  GGN_CLIENT_BOX_H
+#endif  //  ZEN_CLIENT_SIMPLE_BOX_BOX_H
